+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 - [BUGFIX] Fix `view.time_spent` in RUM view events. See [#1596][]
 
+- [FEATURE] Start RUM session on RUM init. See [#1594][]
+
 # 2.5.0 / 08-11-2023
 
 - [BUGFIX] Optimize Session Replay diffing algorithm. See [#1524][]
@@ -561,6 +563,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1524]: https://github.com/DataDog/dd-sdk-ios/pull/1524
 [#1529]: https://github.com/DataDog/dd-sdk-ios/pull/1529
 [#1533]: https://github.com/DataDog/dd-sdk-ios/pull/1533
+[#1594]: https://github.com/DataDog/dd-sdk-ios/pull/1594
 [#1536]: https://github.com/DataDog/dd-sdk-ios/pull/1536
 [#1531]: https://github.com/DataDog/dd-sdk-ios/pull/1531
 [#1596]: https://github.com/DataDog/dd-sdk-ios/pull/1596